Fix zone loading: Heavy objects now placed in rooms, not mobile inventory
- Fixed spawn logic to check object weight and type before trying mobile assignment
- Objects over 100 weight or type Fountain now go to rooms by default
- Eliminated 'Failed to give object to mobile' errors for room fixtures like wells
- Tests still pass, server loads successfully with clean logs

+++ src/core/string_utils.cpp
@@ -23,7 +23,7 @@
 #include <utility>
 
 std::string ellipsis(const std::string_view str, int maxlen) {
-    if (str.length() < maxlen - 3)
+    if (static_cast<int>(str.length()) < maxlen - 3)
         return std::string(str);
 
     std::string result;
